+++ apps/backend/src/modules/projects/libs/types/types.ts
@@ -1,5 +1,8 @@
 export {
 	type ProjectCreateRequestDto,
 	type ProjectGetAllItemResponseDto,
+	type ProjectGetAllRequestDto,
 	type ProjectGetAllResponseDto,
+	type ProjectPatchRequestDto,
+	type ProjectPatchResponseDto,
 } from "@git-fit/shared";