+++ apps/backend/src/modules/projects/project.controller.ts
@@ -8,8 +8,15 @@ import { HTTPCode } from "~/libs/modules/http/http.js";
 import { type Logger } from "~/libs/modules/logger/logger.js";
 
 import { ProjectsApiPath } from "./libs/enums/enums.js";
-import { type ProjectCreateRequestDto } from "./libs/types/types.js";
-import { projectCreateValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
+import {
+	type ProjectCreateRequestDto,
+	type ProjectGetAllRequestDto,
+	type ProjectPatchRequestDto,
+} from "./libs/types/types.js";
+import {
+	projectCreateValidationSchema,
+	projectPatchValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";
 import { type ProjectService } from "./project.service.js";
 
 /**
@@ -59,7 +66,12 @@ class ProjectController extends BaseController {
 		});
 
 		this.addRoute({
-			handler: () => this.findAll(),
+			handler: (options) =>
+				this.findAllByName(
+					options as APIHandlerOptions<{
+						query: ProjectGetAllRequestDto;
+					}>,
+				),
 			method: "GET",
 			path: ProjectsApiPath.ROOT,
 		});
@@ -73,6 +85,21 @@ class ProjectController extends BaseController {
 				),
 			method: "GET",
 			path: ProjectsApiPath.$ID,
+		});
+
+		this.addRoute({
+			handler: (options) =>
+				this.patch(
+					options as APIHandlerOptions<{
+						body: ProjectPatchRequestDto;
+						params: { id: string };
+					}>,
+				),
+			method: "PATCH",
+			path: ProjectsApiPath.$ID,
+			validation: {
+				body: projectPatchValidationSchema,
+			},
 		});
 	}
 
@@ -134,9 +161,15 @@ class ProjectController extends BaseController {
 	 *                		items:
 	 *                  		$ref: "#/components/schemas/Project"
 	 */
-	private async findAll(): Promise<APIHandlerResponse> {
+	private async findAllByName(
+		options: APIHandlerOptions<{
+			query: ProjectGetAllRequestDto;
+		}>,
+	): Promise<APIHandlerResponse> {
+		const { query } = options;
+
 		return {
-			payload: await this.projectService.findAll(),
+			payload: await this.projectService.findAllbyName(query),
 			status: HTTPCode.OK,
 		};
 	}
@@ -172,6 +205,57 @@ class ProjectController extends BaseController {
 	): Promise<APIHandlerResponse> {
 		return {
 			payload: await this.projectService.find(Number(options.params.id)),
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /projects/{id}:
+	 *    patch:
+	 *      description: Update project info
+	 *      parameters:
+	 *        - in: path
+	 *          name: id
+	 *          required: true
+	 *          description: The ID of the project to update
+	 *          schema:
+	 *            type: integer
+	 *      requestBody:
+	 *        description: Project data
+	 *        required: true
+	 *        content:
+	 *          application/json:
+	 *            schema:
+	 *              type: object
+	 *              properties:
+	 *                name:
+	 *                  type: string
+	 *                description:
+	 *                  type: string
+	 *      responses:
+	 *        200:
+	 *          description: Successful operation
+	 *          content:
+	 *            application/json:
+	 *              schema:
+	 *                type: object
+	 *                properties:
+	 *                  message:
+	 *                    type: object
+	 *                    $ref: "#/components/schemas/Project"
+	 */
+
+	private async patch(
+		options: APIHandlerOptions<{
+			body: ProjectPatchRequestDto;
+			params: { id: string };
+		}>,
+	): Promise<APIHandlerResponse> {
+		const projectId = Number(options.params.id);
+
+		return {
+			payload: await this.projectService.patch(projectId, options.body),
 			status: HTTPCode.OK,
 		};
 	}