+++ apps/backend/src/modules/projects/project.repository.ts
@@ -1,6 +1,7 @@
 import { SortType } from "~/libs/enums/enums.js";
 import { type Repository } from "~/libs/types/types.js";
 
+import { type ProjectPatchRequestDto } from "./libs/types/types.js";
 import { ProjectEntity } from "./project.entity.js";
 import { type ProjectModel } from "./project.model.js";
 
@@ -36,10 +37,24 @@ class ProjectRepository implements Repository {
 		return item ? ProjectEntity.initialize(item) : null;
 	}
 
-	public async findAll(): Promise<ProjectEntity[]> {
+	public async findAll(): Promise<{ items: ProjectEntity[] }> {
 		const projects = await this.projectModel
 			.query()
 			.orderBy("created_at", SortType.DESCENDING)
+			.execute();
+
+		return {
+			items: projects.map((project) => ProjectEntity.initialize(project)),
+		};
+	}
+
+	public async findAllbyName(name: string): Promise<ProjectEntity[]> {
+		const projects = await this.projectModel
+
+			.query()
+			.orderBy("created_at", SortType.DESCENDING)
+
+			.whereILike("name", `%${name}%`)
 			.execute();
 
 		return projects.map((project) => ProjectEntity.initialize(project));
@@ -49,6 +64,19 @@ class ProjectRepository implements Repository {
 		const item = await this.projectModel.query().findOne({ name });
 
 		return item ? ProjectEntity.initialize(item) : null;
+	}
+
+	public async patch(
+		id: number,
+		projectData: ProjectPatchRequestDto,
+	): Promise<ProjectEntity> {
+		const { description, name } = projectData;
+
+		const updatedItem = await this.projectModel
+			.query()
+			.patchAndFetchById(id, { description, name });
+
+		return ProjectEntity.initialize(updatedItem);
 	}
 
 	public update(): ReturnType<Repository["update"]> {