+++ apps/frontend/src/pages/access-management/libs/components/group-options/groups-options.tsx
@@ -1,14 +1,23 @@
-import { TableOptions } from "~/libs/components/components.js";
+import { Menu, MenuItem } from "~/libs/components/components.js";
+import { usePopover } from "~/libs/hooks/hooks.js";
+
+import styles from "./styles.module.css";
 
 type Properties = {
 	onEdit: () => void;
 };
 
 const GroupOptions = ({ onEdit }: Properties): React.ReactElement => {
+	const { isOpened, onClose, onOpen } = usePopover();
+
 	return (
-		<TableOptions>
-			<button onClick={onEdit}>Edit</button>
-		</TableOptions>
+		<div className={styles["options-cell"]}>
+			<div className={styles["options-wrapper"]}>
+				<Menu isOpened={isOpened} onClose={onClose} onOpen={onOpen}>
+					<MenuItem iconName="pencil" label="Edit" onClick={onEdit} />
+				</Menu>
+			</div>
+		</div>
 	);
 };