+++ apps/frontend/src/pages/projects/components/components.ts
@@ -1,2 +1,5 @@
 export { ProjectCard } from "./project-card/project-card.js";
 export { ProjectCreateForm } from "./project-create-form/project-create-form.js";
+export { ProjectMenu } from "./project-menu/project-menu.js";
+export { ProjectsSearch } from "./project-search/project-search.js";
+export { ProjectUpdateForm } from "./project-update-form/project-update-form.js";