+++ packages/shared/src/index.ts
@@ -15,6 +15,7 @@ export {
 	changeCase,
 	configureString,
 	formatDate,
+	initDebounce,
 } from "./libs/helpers/helpers.js";
 export { type Config } from "./libs/modules/config/config.js";
 export {
@@ -27,6 +28,8 @@ export {
 } from "./libs/modules/http/http.js";
 export { type Storage } from "./libs/modules/storage/storage.js";
 export {
+	type PaginationQueryParameters,
+	type PaginationResponseDto,
 	type ServerCommonErrorResponse,
 	type ServerErrorDetail,
 	type ServerErrorResponse,
@@ -59,7 +62,11 @@ export {
 	projectCreateValidationSchema,
 	ProjectError,
 	type ProjectGetAllItemResponseDto,
+	type ProjectGetAllRequestDto,
 	type ProjectGetAllResponseDto,
+	type ProjectPatchRequestDto,
+	type ProjectPatchResponseDto,
+	projectPatchValidationSchema,
 	ProjectsApiPath,
 } from "./modules/projects/projects.js";
 export {