+++ packages/shared/src/modules/projects/libs/types/types.ts
@@ -1,3 +1,6 @@
 export { type ProjectCreateRequestDto } from "./project-create-request-dto.type.js";
 export { type ProjectGetAllItemResponseDto } from "./project-get-all-item-response-dto.type.js";
+export { type ProjectGetAllRequestDto } from "./project-get-all-request-dto.type.js";
 export { type ProjectGetAllResponseDto } from "./project-get-all-response-dto.type.js";
+export { type ProjectPatchRequestDto } from "./project-patch-request-dto.type.js";
+export { type ProjectPatchResponseDto } from "./project-patch-response-dto.type.js";