+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.7.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.6.0...shared-v1.7.0) (2024-09-04)
+
+
+### Features
+
+* add only latin validation rule gf-131 ([#173](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/173)) ([ecabee8](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/ecabee8bde62d38d4eaa00886d5765c5bb25b89f))
+
+## [1.6.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.5.0...shared-v1.6.0) (2024-09-04)
+
+
+### Features
+
+* add an ability to edit a project gf-88 ([#137](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/137)) ([92b59e0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/92b59e0f955332412fe52a3fbfcdec570ef0551c))
+
+## [1.5.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.4.0...shared-v1.5.0) (2024-09-04)
+
+
+### Features
+
+* add pagination to the users table gf-82 ([#143](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/143)) ([c663ca0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/c663ca0be448bfd16c85b99c32e1f334199ea03a))
+
+## [1.4.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.3.0...shared-v1.4.0) (2024-09-03)
+
+
+### Features
+
+* Add search to the project list gf-85 ([#114](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/114)) ([44c5ca9](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/44c5ca9d28035094efd5a13a47d65009833611e2))
+
 ## [1.3.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.2.0...shared-v1.3.0) (2024-09-02)